Improve the efficiency of `alignment::distance::levenshtein`

Speed improvement by more than 30%; space usage reduced by half.

## Benchmarks 

Before:

```
test bench_levenshtein_dist_diverse_str              ... bench:  71,196,664 ns/iter (+/- 1,809,200)
test bench_levenshtein_dist_equal_str                ... bench:  71,264,536 ns/iter (+/- 5,608,108)
test bench_levenshtein_dist_worst_str                ... bench:  73,698,633 ns/iter (+/- 7,015,925)
```

After:

```
test bench_levenshtein_dist_diverse_str              ... bench:  49,305,844 ns/iter (+/- 5,275,531)
test bench_levenshtein_dist_equal_str                ... bench:  48,909,164 ns/iter (+/- 4,161,517)
test bench_levenshtein_dist_worst_str                ... bench:  48,370,914 ns/iter (+/- 4,088,614)
```