@pmarks 2020-09-17
@TianyiShi2001 cool  See this example on the compiler explorer: https://godbolt.org/z/7rqadE

The compiler is able to elide the bounds check when iterating from 0..m, but not when iterating from 1..=m, and subtracting 1 from the index.  It's a bit annoying/unreliable to have to write code a particular way to get the perf improvement.  Not sure what is best, using unsafe, or tweaking the safe code to coax the compiler to elide the bounds check.