@tedil 2020-09-18
> @TianyiShi2001 cool See this example on the compiler explorer: https://godbolt.org/z/7rqadE
> 
> The compiler is able to elide the bounds check when iterating from 0..m, but not when iterating from 1..=m, and subtracting 1 from the index. It's a bit annoying/unreliable to have to write code a particular way to get the perf improvement. Not sure what is best, using unsafe, or tweaking the safe code to coax the compiler to elide the bounds check.

The fewer `unsafe`s, the better! Perhaps we could also try to rewrite the code in terms of iterators, since the compiler is pretty good at reasoning about them. (i.e. `a.iter().for_each(|&a_i| {v += a_i as i32});` results in the exact same code)