@pmarks 2020-09-18
@TianyiShi2001 I agree with your logic here.  The `unsafe` usages in this code are trivially verifiable, so they don't introduce too much risk.  I do we should continue to keep unsafe code very localized & use it sparingly (and also look for safe abstractions that guarantee high performance _and_ keep the code readable)