@TianyiShi2001 2020-09-18
I also found that bound checks are not elided when looping over a naïve matrix made of `Vec<Vec<T>>`. See: https://gist.github.com/TianyiShi2001/7f83854b91a94f3eaf3145084db6d627