@veldsla 2020-10-07
I hate to be that guy that's always complaining, but I just don't like unsafe. And I don't like the recent trend of sticking in get_unchecked everywhere just to gain a few percent. Maybe in this case is it trivially verifiable, but why would I want to? I really like the conclusion at the end of this thread https://users.rust-lang.org/t/why-is-this-rust-code-slower-than-c/49564/66

> using cheap tricks like unsafe get_unchecked() doesn't normally do much other than open you up to bugs. Thousands of man-hours have been poured into LLVM's optimiser and it'll automatically apply these optimisations where it knows it's safe to do so...

So I tried putting my...erm..time, where my mouth is. Looking at the latest version of the levenshtein function in this PR it is already pretty clean. Without giving any thought to the algorithm and the correctness of the implementation I simply translated it to iterators:
```rust
pub fn lev_safe(alpha: &[u8], beta: &[u8]) -> usize {
    let m = alpha.len();
    let mut dp_matrix: Vec<_> = (0..(m+1)).collect();

    beta.iter().enumerate().for_each(|(j, b)| {
        let mut s_diag = j;
        let mut s_above = j + 1;
        dp_matrix.iter_mut().skip(1)
            .zip(alpha)
            .for_each(|(dp, a)| {
                s_above = min(
                    s_diag + if a == b { 0 } else { 1 },
                    min(s_above + 1, *dp + 1),
                );
                s_diag = *dp;
                *dp = s_above;
            });
    });

    dp_matrix[m]
}
```
I ran a Criterion benchmark and the safe version is on my hardware even faster (forgive me my misspelling of levenshtein):
![violin_lv](https://user-images.githubusercontent.com/592718/95322135-5346c780-089c-11eb-976a-26c26357ab8e.png)
The only 'trick' I had to use was swapping the `alpha` and the `dp_matrix` iterator while zipping. Most likely putting skip(1) in the zip prevented some optimization and turned out to be about 10%slower.

Oh, this version returns usize, seems weird that a function that compares slices should return u32 for the distance.